test case of misbehaviour of rule PN on top of simulator.
